bug: Do not run clippy on beta toolchain
As observed in #24, only invoke Clippy for the `Stable` and `Nightly` release channels on Travis CI.

Unfortunately, versions of Clippy vary across channels. Only run Clippy against the `Stable` channel, as well as `Nightly` which is allowed to fail. The `Beta` release channel is still tested against, but only for crate tests and not code linting.

Closes #25